requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
absl-py==2.3.1
    # via mediapipe
attrs==25.4.0
    # via mediapipe
cffi==2.0.0
    # via sounddevice
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.20.0
    # via torch
flatbuffers==25.9.23
    # via mediapipe
fonttools==4.60.1
    # via matplotlib
fsspec==2025.10.0
    # via torch
jax==0.4.38
    # via mediapipe
jaxlib==0.4.38
    # via
    #   jax
    #   mediapipe
jinja2==3.1.6
    # via torch
joblib==1.5.2
    # via scikit-learn
kiwisolver==1.4.9
    # via matplotlib
markupsafe==3.0.3
    # via jinja2
matplotlib==3.8.4
    # via
    #   -r requirements.in
    #   mediapipe
mediapipe==0.10.14
    # via -r requirements.in
ml-dtypes==0.5.3
    # via
    #   jax
    #   jaxlib
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   jax
    #   jaxlib
    #   matplotlib
    #   mediapipe
    #   ml-dtypes
    #   opencv-contrib-python
    #   pandas
    #   scikit-learn
    #   scipy
opencv-contrib-python==4.8.1.78
    # via
    #   -r requirements.in
    #   mediapipe
opt-einsum==3.4.0
    # via jax
packaging==25.0
    # via matplotlib
pandas==2.3.3
    # via -r requirements.in
pillow==12.0.0
    # via matplotlib
protobuf==4.25.8
    # via mediapipe
pycparser==2.23
    # via cffi
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via -r requirements.in
scikit-learn==1.4.2
    # via -r requirements.in
scipy==1.15.3
    # via
    #   jax
    #   jaxlib
    #   scikit-learn
six==1.17.0
    # via python-dateutil
sounddevice==0.5.3
    # via mediapipe
sympy==1.14.0
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
torch==2.2.2
    # via -r requirements.in
tqdm==4.67.1
    # via -r requirements.in
typing-extensions==4.15.0
    # via torch
tzdata==2025.2
    # via pandas
